🧪 [testing improvement] Add missing tests for EnergyState.__post_init__ validation
